Revision 2

+++ routing.py
@@ -4,21 +4,28 @@
 # 3/26/18
 ####################################
 
-from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
+from flask import (Flask,
+                   render_template,
+                   request,
+                   redirect,
+                   url_for,
+                   flash,
+                   jsonify,
+                   session as login_session,
+                   g,
+                   make_response)
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from db_setup import Base, Playlists, Songs
-import random, string
-from flask import session as login_session
+import random
+import string
 import json
 import requests
-from oauth2client.client import flow_from_clientsecrets
-from oauth2client.client import FlowExchangeError
+from oauth2client.client import flow_from_clientsecrets, FlowExchangeError
 import httplib2
-from flask import g
-from flask import make_response
 
-CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
+CLIENT_ID = json.loads(open(
+    'client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Playlist App"
 
 
@@ -31,12 +38,13 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-login="login"
+login = "login"
 app.jinja_env.globals.update(login="login")
-logout=""
+logout = ""
 ######################################
 # -------- Routing Functions ---------
 ######################################
+
 
 # ------------ Main page - View all playlists -------------
 @app.route('/')
@@ -48,23 +56,25 @@ def mainPage():
         user_string = "Logged in as %s" % login_session['username']
     return render_template("main.html", array=array)
 
+
 # ---------------- Login page ----------------------
-@app.route('/login', methods=['GET','POST'] )
+@app.route('/login', methods=['GET', 'POST'])
 def login():
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
                     for x in range(32))
     login_session['state'] = state
     return render_template('login.html', STATE=state)
 
+
 # ------------- gconnect -----------------------
-@app.route('/gconnect', methods=['GET','POST'] )
+@app.route('/gconnect', methods=['GET', 'POST'])
 def gconnect():
-   # Validate state token
+    # Validate state token
     a = request.args.get('state')
     b = login_session['state']
     print(a)
     print(b)
-    if  a != b:
+    if a != b:
         response = make_response(json.dumps('Invalid state parameter.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
@@ -102,8 +112,8 @@ def gconnect():
         response = make_response((error), 500)
         response.headers['Content-Type'] = 'application/json'
         return response
-    except:
-      print("no content['error']")
+    except error:
+        print("no content['error']")
 
     # Verify that the access token is used for the intended user.
     gplus_id = int(credentials.id_token['sub'])
@@ -120,11 +130,11 @@ def gconnect():
         return response
 
     # Verify that the access token is valid for this app.
-    issued_to = content['issued_to'] 
+    issued_to = content['issued_to']
     print ("issued to :")
     print(issued_to)
     print("CLIENT_ID ")
-    print(CLIENT_ID) 
+    print(CLIENT_ID)
     if issued_to != CLIENT_ID:
         response = make_response(
             json.dumps("Token's client ID does not match app's."), 401)
@@ -135,8 +145,8 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(json.dumps(
+            'Current user is already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -162,19 +172,22 @@ def gconnect():
     output += '<h1>Welcome</h1>'
     return output
 
+
 # --------------- gdisconnect ----------------
 @app.route('/gdisconnect')
 def gdisconnect():
     access_token = login_session.get('access_token')
     if access_token is None:
         print ('Access Token is None')
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(json.dumps(
+            'Current user not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
     print ('In gdisconnect access token is %s', access_token)
     print ('User name is: ')
     print (login_session['username'])
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % (
+        login_session['access_token'])
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print ('result is ')
@@ -185,42 +198,50 @@ def gdisconnect():
         del login_session['username']
         del login_session['email']
         del login_session['picture']
-        #response = make_response(json.dumps('Successfully disconnected.'), 200)
-        #response.headers['Content-Type'] = 'application/json'
         app.jinja_env.globals.update(login="login")
         app.jinja_env.globals.update(logout="")
+
         return redirect(url_for('mainPage'))
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(json.dumps(
+            'Failed to revoke token for given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
+
 
 # ------------ View a playlist -------------
 @app.route('/<int:playlist_id>')
 def viewPlaylist(playlist_id):
-    playlist = session.query(Playlists).filter_by(id = playlist_id).one()
-    songs = session.query(Songs).filter_by(playlistId = playlist_id)
-    return render_template("playlist.html", songs=songs, playlist_id=playlist_id, playlist=playlist)
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    songs = session.query(Songs).filter_by(playlistId=playlist_id)
+    return render_template("playlist.html", songs=songs,
+                           playlist_id=playlist_id, playlist=playlist)
+
 
 # ------------ Add a playlist --------------
-@app.route('/add/', methods=['GET','POST'])
+@app.route('/add/', methods=['GET', 'POST'])
 def addPlaylist():
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
-        newPlaylist = Playlists(name = request.form['name'])
+        newPlaylist = Playlists(name=request.form['name'],
+                                user_id=login_session['email'])
         session.add(newPlaylist)
         session.commit()
         return redirect(url_for('mainPage'))
     else:
         return render_template("add_playlist.html")
 
+
 # ------------- Edit a playlist -------------
-@app.route('/edit/<int:playlist_id>', methods=['GET','POST'])
+@app.route('/edit/<int:playlist_id>', methods=['GET', 'POST'])
 def editPlaylist(playlist_id):
     if 'username' not in login_session:
         return redirect('/login')
-    playlist = session.query(Playlists).filter_by(id = playlist_id).one()
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    if login_session['email'] != playlist.user_id:
+        flash('You do not own this playlist')
+        return render_template('no_auth.html')
     if request.method == 'POST':
         newName = request.form['name']
         playlist.name = newName
@@ -228,12 +249,16 @@ def editPlaylist(playlist_id):
     else:
         return render_template("edit_playlist.html", playlist=playlist)
 
+
 # -------------- Delete a playlist -------------
-@app.route('/delete/<int:playlist_id>', methods=['GET','POST'])
+@app.route('/delete/<int:playlist_id>', methods=['GET', 'POST'])
 def deletePlaylist(playlist_id):
     if 'username' not in login_session:
         return redirect('/login')
-    playlist = session.query(Playlists).filter_by(id = playlist_id).one()
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    if login_session['email'] != playlist.user_id:
+        flash('You do not own this playlist')
+        return render_template('no_auth.html')
     if request.method == 'POST':
         session.delete(playlist)
         session.commit()
@@ -242,36 +267,53 @@ def deletePlaylist(playlist_id):
     else:
         return render_template("delete_playlist.html", playlist=playlist)
 
+
 # -------------- View a song --------------------
 @app.route('/<int:playlist_id>/<int:song_id>')
 def viewSong(playlist_id, song_id):
-    song = session.query(Songs).filter_by(playlistId = playlist_id, id=song_id).one()
-    return render_template("view_song.html", song=song, playlist_id=playlist_id, song_id=song_id)
+    song = session.query(Songs).filter_by(playlistId=playlist_id,
+                                          id=song_id).one()
+    return render_template("view_song.html", song=song,
+                           playlist_id=playlist_id, song_id=song_id)
+
 
 # --------------- Add a song -----------------
-@app.route('/add/<int:playlist_id>', methods=['GET','POST'] )
+@app.route('/add/<int:playlist_id>', methods=['GET', 'POST'])
 def addSong(playlist_id):
     if 'username' not in login_session:
         return redirect('/login')
-    playlist = session.query(Playlists).filter_by(id = playlist_id).one()
-
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    if login_session['email'] != playlist.user_id:
+        flash('You do not own this playlist')
+        return render_template('no_auth.html')
     if request.method == 'POST':
 
-        newSong = Songs(title = request.form['title'], artist = request.form['artist'], playlistId = playlist_id, album = request.form['album'], 
-            duration = request.form['duration'], artwork = request.form['artwork'], youtubeId = request.form['youtubeId'])
+        newSong = Songs(title=request.form['title'],
+                        artist=request.form['artist'],
+                        playlistId=playlist_id,
+                        album=request.form['album'],
+                        duration=request.form['duration'],
+                        artwork=request.form['artwork'],
+                        youtubeId=request.form['youtubeId'])
 
         session.add(newSong)
         session.commit()
         return redirect(url_for('viewPlaylist', playlist_id=playlist_id))
     else:
-        return render_template("add_song.html", playlist_id=playlist_id, playlist=playlist)
+        return render_template("add_song.html",
+                               playlist_id=playlist_id, playlist=playlist)
+
 
 # --------------- Edit a song ----------------
-@app.route('/edit/<int:playlist_id>/<int:song_id>', methods=['GET','POST'])
+@app.route('/edit/<int:playlist_id>/<int:song_id>', methods=['GET', 'POST'])
 def editSong(playlist_id, song_id):
     if 'username' not in login_session:
         return redirect('/login')
-    song = session.query(Songs).filter_by(id = song_id).one()
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    if login_session['email'] != playlist.user_id:
+        flash('You do not own this playlist')
+        return render_template('no_auth.html')
+    song = session.query(Songs).filter_by(id=song_id).one()
     if request.method == 'POST':
         song.title = request.form['title']
         song.artist = request.form['artist']
@@ -283,39 +325,50 @@ def editSong(playlist_id, song_id):
 
         return redirect(url_for('viewPlaylist', playlist_id=playlist_id))
     else:
-        return render_template("edit_song.html", playlist_id=playlist_id, song=song)
+        return render_template("edit_song.html", playlist_id=playlist_id,
+                               song=song)
+
 
 # --------------- Delete a song -----------------
-@app.route('/delete/<int:playlist_id>/<int:song_id>', methods=['GET','POST'])
+@app.route('/delete/<int:playlist_id>/<int:song_id>', methods=['GET', 'POST'])
 def deleteSong(playlist_id, song_id):
     if 'username' not in login_session:
         return redirect('/login')
-    playlist = session.query(Playlists).filter_by(id = playlist_id).one()
-    songToDelete = session.query(Songs).filter_by(id = song_id).one()
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    if login_session['email'] != playlist.user_id:
+        flash('You do not own this playlist')
+        return render_template('no_auth.html')
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    songToDelete = session.query(Songs).filter_by(id=song_id).one()
     if request.method == 'POST':
         session.delete(songToDelete)
         session.commit()
         session.flush()
         return redirect(url_for('viewPlaylist', playlist_id=playlist_id))
     else:
-        return render_template("delete_song.html", playlist_id=playlist_id, song=songToDelete, playlist=playlist)
+        return render_template("delete_song.html", playlist_id=playlist_id,
+                               song=songToDelete, playlist=playlist)
+
 
 # ---------------- JSON endpoints ---------------
 @app.route('/json/mainpage')
 def jsonMainPage():
     array = session.query(Playlists).order_by(asc(Playlists.name))
-    return jsonify(playlists= [p.serialize for p in array])
+    return jsonify(playlists=[p.serialize for p in array])
+
 
 @app.route('/json/playlist/<int:playlist_id>')
 def jsonViewPlaylist(playlist_id):
-    playlist = session.query(Playlists).filter_by(id = playlist_id).one()
-    songs = session.query(Songs).filter_by(playlistId = playlist_id)
-    return jsonify(songs = [s.serialize for s in songs])
+    playlist = session.query(Playlists).filter_by(id=playlist_id).one()
+    songs = session.query(Songs).filter_by(playlistId=playlist_id)
+    return jsonify(songs=[s.serialize for s in songs])
+
 
 @app.route('/json/song/<int:playlist_id>/<int:song_id>')
 def jsonViewSong(playlist_id, song_id):
-    song = session.query(Songs).filter_by(playlistId = playlist_id, id=song_id).one()
-    return jsonify(song = song.serialize)
+    song = session.query(Songs).filter_by(
+        playlistId=playlist_id, id=song_id).one()
+    return jsonify(song=song.serialize)
 
 
 ##########################################
@@ -324,4 +377,4 @@ def jsonViewSong(playlist_id, song_id):
 if __name__ == '__main__':
         app.secret_key = 'secret_key'
         app.debug = True
-        app.run(host = '0.0.0.0', port = 8000)
+        app.run(host='0.0.0.0', port=8000)
